QR Generator/requirements.txt
pandas
numpy
qrcode
pillow
configparser
openpyxl